add type declare file
