+++ src/liff.d.ts
@@ -1,0 +1,10 @@
+import { ShareTargetPickerResult } from "./types";
+
+declare module "@line/liff" {
+  interface Liff {
+    $deepdive: {
+      shareMessages: () => Promise<ShareTargetPickerResult | void>;
+      readQrCode: () => Promise<void>;
+    };
+  }
+}
